+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "vscode-tui-language-support" extension will be docum
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2026-08-20
+
+### Added
+
+- Basic formatting support for nullable types and table literals
+- Formatter configuration options
+
 ## [0.3.1] - 2026-08-19
 
 ### Fixed